logica

+++ CrediMujer/Logica/Credito.cs
@@ -4,8 +4,8 @@ namespace CrediMujer.Logica
 {
     public class Credito
     {
-        public float Importe { get; set; }
+        public double Importe { get; set; }
         public int Plazo { get; set; }
-        public float PagoMensual { get; set; }
+        public double PagoMensual { get; set; }
     }
 }
